Move feature flags constants to the constants pkg
Fixes #34

The feature flags contants where part of the "config" package and
the "config" package uses the "discovery" package.  Such a setup
prevents the discovery package from using feature flags as it would
cause a circular dependency.

By moving the feature flag constants to the "constants" package, the
"discovery" package can now use it.
